[16.0][IMP] report_xml: Added xml_extension
Adds a way to choose the file extension for your XML file, by adding the flexibility of appending more choices.
For example, instead of getting `.xml`, this allows to have the report download as `.svg` or even good old `.html` if needed.
This is also very useful for some localization purposes, where some scarce software uses uncommon file extensions for their XML files, like `.ffdata` for specific accounting reports here in Lithuania.

The change is not breaking, as we set the default to be `.xml` like it was by default anyway.